+++ moabb/datasets/liu2024.py
@@ -90,6 +90,7 @@ class Liu2024(BaseDataset):
 
     """
 
+    nemar_id = "nm000158"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=500.0,