+++ moabb/datasets/romani_bf2025_erp.py
@@ -127,6 +127,7 @@ class RomaniBF2025ERP(BaseDataset):
 
     """
 
+    nemar_id = "nm000272"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=250.0,