+++ moabb/datasets/ssvep_exo.py
@@ -71,6 +71,7 @@ class Kalunga2016(BaseDataset):
 
     """
 
+    nemar_id = "nm000123"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=256.0,